refactor: consolidate tracers module types
