+++ openfisca_core/tracers/flat_trace.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 import numpy
 
+from openfisca_core import types as t
 from openfisca_core.indexed_enums import EnumArray
-
-from . import types as t
 
 
 class FlatTrace:
