+++ openfisca_core/tracers/simple_tracer.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from . import types as t
+from openfisca_core import types as t
 
 
 class SimpleTracer:
